refactor: テストコードの共通fixtureをconftest.pyに集約
各テストファイルで重複していたCLIPモデル/プロセッサのモック設定を
tests/conftest.pyに一元化し、メンテナンス性とデバッグ性を向上させる。
あわせてGitHub Actionsのテスト実行設定を改善し、出力をリアルタイム表示
するよう変更。
