+++ tests/analyzers/test_clip_model_manager.py
@@ -7,121 +7,10 @@
 4. 高速実行（約2-5秒） - 重いモデルロードなし
 """
 
-from collections.abc import Generator
-from typing import Any
-from unittest.mock import MagicMock, patch
+from unittest.mock import MagicMock
 
-import pytest
-import torch
 
 from src.analyzers.clip_model_manager import CLIPModelManager
-
-
-@pytest.fixture(autouse=True)
-def mock_clip_model() -> Generator[Any, Any, Any]:
-    """700MBの重みロードを回避するためのCLIPモデルのモック.
-
-    このfixtureは本物のCLIPモデルを以下のモックに置き換えます：
-    - 決定論的テストのために固定されたlogit値を返す
-    - GPU/CPU切り替えのための.to(device)呼び出しをサポート
-    - 512次元のCLIP特徴ベクトルを返す（実際のモデルと同じ形状）
-    - バッチサイズに応じた形状を返す（動的対応）
-    """
-    with patch("transformers.CLIPModel.from_pretrained") as mock:
-        model = MagicMock()
-
-        # get_text_features用のモック（テキスト埋め込み）: (batch_size, 512)
-        def mock_get_text_features(**kwargs: object) -> torch.Tensor:
-            # input_idsからバッチサイズを取得
-            inputs = kwargs.get("input_ids")
-            if inputs is not None and isinstance(inputs, torch.Tensor):
-                batch_size = inputs.shape[0]
-            else:
-                batch_size = 1
-            # 正規化された固定ベクトルを返す
-            return torch.ones(batch_size, 512) / torch.sqrt(torch.tensor(512.0))
-
-        # get_image_features用のモック（画像埋め込み）: (batch_size, 512)
-        def mock_get_image_features(**kwargs: object) -> torch.Tensor:
-            # pixel_valuesからバッチサイズを取得
-            inputs = kwargs.get("pixel_values")
-            if inputs is not None and isinstance(inputs, torch.Tensor):
-                batch_size = inputs.shape[0]
-            else:
-                batch_size = 1
-            # 正規化された固定ベクトルを返す
-            return torch.ones(batch_size, 512) / torch.sqrt(torch.tensor(512.0))
-
-        # メソッドをモック
-        model.get_text_features = MagicMock(side_effect=mock_get_text_features)
-        model.get_image_features = MagicMock(side_effect=mock_get_image_features)
-        model.to = MagicMock(return_value=model)
-        model.eval = MagicMock()
-
-        mock.return_value = model
-        yield
-
-
-@pytest.fixture(autouse=True)
-def mock_clip_processor() -> Generator[Any, Any, Any]:
-    """トークナイザと特徴抽出器のロードを回避するためのCLIPプロセッサのモック.
-
-    このfixtureは本物のCLIPプロセッサを以下のモックに置き換えます：
-    - テキストと画像のために固定されたテンソル形状を返す
-    - GPU/CPU切り替えのための.to(device)呼び出しをサポート
-    - バッチサイズに応じた形状を返す（動的対応）
-    """
-    with patch("transformers.CLIPProcessor.from_pretrained") as mock:
-        # processorは呼び出し可能で、辞書のようなオブジェクトを返す
-        processor = MagicMock()
-
-        def mock_processor(**kwargs: object) -> MagicMock:
-            # imagesからバッチサイズを取得
-            images = kwargs.get("images")
-            if images is not None:
-                if isinstance(images, list):
-                    batch_size = len(images)
-                else:
-                    batch_size = 1
-            else:
-                batch_size = 1
-
-            # 呼び出し時に返す辞書のようなオブジェクト
-            input_ids = torch.tensor([[1, 2, 3]])
-            pixel_values = torch.ones(batch_size, 3, 224, 224) * 0.5
-            attention_mask = torch.tensor([[1, 1, 1]])
-
-            # MagicMockを作成して属性を設定
-            result_obj = MagicMock()
-            result_obj.input_ids = input_ids
-            result_obj.pixel_values = pixel_values
-            result_obj.attention_mask = attention_mask
-
-            # 辞書のように振る舞うための__getitem__を実装
-            def getitem(_self: MagicMock, key: str) -> torch.Tensor:
-                if key == "input_ids":
-                    return input_ids
-                elif key == "pixel_values":
-                    return pixel_values
-                elif key == "attention_mask":
-                    return attention_mask
-                else:
-                    raise KeyError(key)
-
-            # 束縛されたメソッドをMagicMockに設定
-            import types
-
-            result_obj.__getitem__ = types.MethodType(getitem, result_obj)
-
-            # .to()メソッドをサポート（自分自身を返す）
-            result_obj.to = MagicMock(return_value=result_obj)
-
-            return result_obj
-
-        processor.side_effect = mock_processor
-
-        mock.return_value = processor
-        yield
 
 
 def test_initialization_with_default_parameters() -> None:
